+++ common/db/Cargo.toml
@@ -14,9 +14,9 @@ all-features = true
 rustdoc-args = ["--cfg", "docsrs"]
 
 [dependencies]
-redb = { version = "1", default-features = false, optional = true }
+parity-db = { version = "0.4", default-features = false, optional = true }
 rocksdb = { version = "0.21", default-features = false, features = ["lz4"], optional = true }
 
 [features]
-redb = ["dep:redb"]
+parity-db = ["dep:parity-db"]
 rocksdb = ["dep:rocksdb"]
