+++ orchestration/coordinator/Dockerfile
@@ -46,7 +46,7 @@ RUN --mount=type=cache,target=/root/.cargo \
   --mount=type=cache,target=/usr/local/cargo/git \
   --mount=type=cache,target=/serai/target \
   mkdir /serai/bin && \
-  cargo build -p serai-coordinator --features redb && \
+  cargo build -p serai-coordinator --features parity-db && \
   mv /serai/target/debug/serai-coordinator /serai/bin
 FROM debian:bookworm-slim as image
