+++ orchestration/message-queue/Dockerfile
@@ -46,7 +46,7 @@ RUN --mount=type=cache,target=/root/.cargo \
   --mount=type=cache,target=/usr/local/cargo/git \
   --mount=type=cache,target=/serai/target \
   mkdir /serai/bin && \
-  cargo build --features redb -p serai-message-queue && \
+  cargo build --features parity-db -p serai-message-queue && \
   mv /serai/target/debug/serai-message-queue /serai/bin
 FROM debian:bookworm-slim as image
