+++ orchestration/message-queue/Dockerfile.message-queue
@@ -1,2 +1,2 @@
-  cargo build --features redb -p serai-message-queue && \
+  cargo build --features parity-db -p serai-message-queue && \
   mv /serai/target/debug/serai-message-queue /serai/bin
